# Conflicts:
#	src/main/java/MSACHAT/backend/BackendApplication.java
#	src/main/java/MSACHAT/backend/controller/CommentController.java
#	src/main/java/MSACHAT/backend/dto/CommentDto.java
#	src/main/java/MSACHAT/backend/dto/PostDto.java
#	src/main/java/MSACHAT/backend/entity/CommentEntity.java
#	src/main/java/MSACHAT/backend/entity/PostEntity.java
#	src/main/java/MSACHAT/backend/repository/CommentRepository.java
#	src/main/java/MSACHAT/backend/repository/PostRepository.java
#	src/main/java/MSACHAT/backend/service/impl/PostServiceImpl.java
#	target/classes/MSACHAT/backend/BackendApplication.class

+++ src/main/java/MSACHAT/backend/service/impl/CustomUserDetailsService.java
@@ -22,20 +22,19 @@ public class CustomUserDetailsService implements UserDetailsService {
     public UserDetails loadUserByUsername(String usernameOrEmail) throws UsernameNotFoundException {
 
         UserEntity userEntity = userRepository.findByUsernameOrEmail(usernameOrEmail, usernameOrEmail)
-                .orElseThrow(() -> new UsernameNotFoundException("User not exists by Username or Email"));
+                .orElseThrow(() -> new UsernameNotFoundException(
+                        "User not exists by Username or Email"));
 
-        Long userId = userEntity.getId();
+        Integer userId = userEntity.getId();
 
         Set<GrantedAuthority> authorities = userEntity.getRoles().stream()
                 .map((role) -> new SimpleGrantedAuthority(role.getName()))
                 .collect(Collectors.toSet());
 
-
         return new org.springframework.security.core.userdetails.User(
 
                 usernameOrEmail,
                 userEntity.getPassword(),
-                authorities
-        );
+                authorities);
     }
 }
